Move settings save logic to SettingsDialog component
